backend/requirements.txt
# Finance Tracker Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# SimpleFIN (uses httpx for API calls)
# No additional package needed - SimpleFIN uses REST API

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Scheduling
apscheduler>=3.10.4

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Date/Time
python-dateutil>=2.8.2

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# CORS for PWA
starlette>=0.35.0
